Added EIO and CIO connections to the U3
The U3-LV and U3-HV connection labels list now have all the
connections on the device, not just the screw terminals. Clutter
isn't too bad. Didn't add the corresponding Modbus addresses to
the modbus/ directory. It's not clear if there is a demand for
that directory until the ability to make your own register files.